Move correlation context and trace context to new namespace
